+++ docs/changes.md
@@ -67,6 +67,21 @@ multi, thread = require("multi"):init{print=true}
 GLOBAL, THREAD = require("multi.integration.lanesManager"):init()
 ```
 
+## Added New Integration: **effilManager**
+
+Another option for multithreading support, works just like all the other threading integrations, but uses the internals of effil and it's unique features.
+- Refer to this [doc](https://www.lua.org/wshop18/Kupriyanov.pdf) to read more about it.
+- Project github [page](https://github.com/effil/effil/tree/master).
+
+```lua
+package.path = "?/init.lua;?.lua;"..package.path
+
+local multi, thread = require("multi"):init({print=true, warn=true, error=true})
+local THREAD, GLOBAL = require("multi.integration.effilManager"):init()
+
+-- Code as you would
+```
+
 ## Added New Integration: **priorityManager**
 
 Allows the user to have multi auto set priorities (Requires chronos). Also adds the functionality to create your own runners (multi:mainloop(), multi:umanager()) that you can set using the priority manager. Even if you do not have `chronos` installed all other features will still work!